Move encryption into the config folder
