+++ src/backend/configManager/configManager.ts
@@ -1,4 +1,4 @@
-import { decrypt, encrypt } from '@/backend/modules/encryption/crypto';
+import { decrypt, encrypt } from '@/backend/configManager/encryption/crypto';
 import { CompanyTypes } from 'israeli-bank-scrapers-core';
 import { promises as fs, existsSync } from 'fs';
 import { Credentials } from '../export/outputVendors/googleSheets/googleAuth';
