+++ src/backend/configManager/encryption/keytar.js
