+++ src/backend/configManager/encryption/salt.js
